backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
alembic>=1.7.3
pytz>=2021.1
PyPDF2>=2.10.5
python-docx>=0.8.11
spacy>=3.1.3
scikit-learn>=1.0.0
bs4>=0.0.1
requests>=2.26.0
celery>=5.1.2
redis>=4.0.2
pandas>=1.3.3
